Improvements

+++ include/gname.h
@@ -133,7 +133,7 @@ typedef struct
     u8 pad1[2];
 } UnkStruct2;
 
-extern void* func_80142274(void* arg0, s32* arg1, u8 arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6, s32 arg7);
+extern void* func_80142274(void* arg0, s32* arg1, s32 arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6, s32 arg7);
 
 extern s32 D_8014F840;
 extern s32 g_name_pixel_width;
